+++ src/lib/util/globalInteractions.ts
@@ -6,7 +6,7 @@ import { buyBingoTicketCommand } from '../../mahoji/commands/bingo';
 import { autoContract } from '../../mahoji/lib/abstracted_commands/farmingContractCommand';
 import { shootingStarsCommand, starCache } from '../../mahoji/lib/abstracted_commands/shootingStarsCommand';
 import { Cooldowns } from '../../mahoji/lib/Cooldowns';
-import { ClueTier } from '../clues/clueTiers';
+import { ClueTier, ClueTiers } from '../clues/clueTiers';
 import { PerkTier } from '../constants';
 import { prisma } from '../settings/prisma';
 import { runCommand } from '../settings/settings';
@@ -228,6 +228,17 @@ async function repeatTripHandler(user: MUser, interaction: ButtonInteraction) {
 	const matchingActivity = trips.find(i => i.type === split[2]);
 	if (!matchingActivity) return repeatTrip(interaction, trips[0]);
 	return repeatTrip(interaction, matchingActivity);
+}
+
+export function buildClueButtons(loot: Bank | null, perkTier: number) {
+	const components: ButtonBuilder[] = [];
+	if (loot && perkTier > PerkTier.One) {
+		const clueReceived = ClueTiers.filter(tier => loot.amount(tier.scrollID) > 0);
+		if (clueReceived.length > 0) {
+			clueReceived.map(clue => components.push(makeDoClueButton(clue)));
+		}
+	}
+	return components;
 }
 
 export async function interactionHook(interaction: Interaction) {
